Add backend function (upload video to cloud, delete video, qiniu hls/screenshot callback, get video types) and fix some problems.

+++ backend/model/video_action.go
@@ -7,9 +7,9 @@ import (
 
 type VideoAction struct {
 	// Id, VideoId, UserId, ActionTime, UnActionTime
-	Id           int64        `default:"0"`
-	VideoId      int          `default:"0"`
-	UserId       int          `default:"0"`
+	Id           uint         `default:"0"`
+	VideoId      uint         `default:"0"`
+	UserId       uint         `default:"0"`
 	ActionTime   time.Time    `default:"0000-00-00 00:00:00"`
 	UnActionTime sql.NullTime `default:"0000-00-00 00:00:00"`
 }
