+++ backend/model/video_history.go
@@ -6,8 +6,8 @@ import (
 
 type VideoHistory struct {
 	// Id, VideoId, UserId, RecordTime
-	Id         int64     `default:"0"`
-	VideoId    int       `default:"0"`
-	UserId     int       `default:"0"`
+	Id         uint      `default:"0"`
+	VideoId    uint      `default:"0"`
+	UserId     uint      `default:"0"`
 	RecordTime time.Time `default:"0000-00-00 00:00:00"`
 }
